Move functions to tools. Start working on script to make tables.
